refactor(app): move boot and subscription wiring into app/lifecycle.rs
Extract the iced runtime glue (Snapdash::boot and Snapdash::subscription)
from snapdash.rs into a dedicated app/lifecycle.rs sibling, kept as
another impl Snapdash block so lib.rs can keep referring to
Snapdash::boot / Snapdash::subscription unchanged.

snapdash.rs now holds the message-handling core; lifecycle owns the
subscription wiring (window/redraw/move events, HA stream, keyboard
focus, hourly update check) and the boot Task fan-out.
